moving main location
