+++ src/test/java/br/com/gymback/application/TestGymbackApplication.java
@@ -1,6 +1,7 @@
 /* Under MIT License (C)2024 */
 package br.com.gymback.application;
 
+import br.com.gymback.GymBackApplication;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
